# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FeedbackHub API
  version: "1.0.0"
  description: |
    REST API for FeedbackHub, a self-hostable feedback & survey widget platform.
    Administrative endpoints require a Bearer access token. Widget-facing endpoints
    (under /api/v1/public and POST /api/v1/responses) are authenticated with a
    project public key instead of a user session.
servers:
  - url: http://localhost:4000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    projectKey:
      type: apiKey
      in: header
      name: X-Project-Key
...
paths:
...
  /projects:
    get:
      summary: List projects, optionally filtered by organisationId
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: List of projects }
    post:
      summary: Create a project within an organisation
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Project created }
  /projects/{id}:
    get:
      summary: Get a project
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Project }
    patch:
      summary: Update a project
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Updated project }
    delete:
      summary: Soft-delete a project
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
...
  /responses:
    get:
      summary: List responses for a project (paginated, filterable)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: projectId, in: query, required: true, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses:
        "200": { description: Paginated responses }
    post:
      summary: Submit a response (public; authenticated via project public key)
      security: [{ projectKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectKey, widgetId]
              properties:
                projectKey: { type: string }
                widgetId: { type: string }
                rating: { type: integer }
                npsScore: { type: integer }
                feedbackText: { type: string }
      responses:
        "201": { description: Response stored }
  /responses/{id}:
    get:
      summary: Get a response with its answers
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Response detail }
  /responses/export/{format}:
    get:
      summary: Export responses as csv or json
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: format, in: path, required: true, schema: { type: string, enum: [csv, json] } }
        - { name: projectId, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Export file }
...
  /webhooks:
    get:
      summary: List webhooks for a project
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: projectId, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: List of webhooks }
    post:
      summary: Create a webhook subscription
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Webhook created (secret returned once) }
...